src/korean-teacher-agent/pyproject.toml
[tool.poetry]
name = "korean-teacher-agent"
version = "0.1.0"
description = ""
authors = ["RRoundTable <ryu071511@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
langchain-openai = ">=0.3.14,<0.4.0"
pydantic = ">=2.11.3,<3.0.0"
langgraph = ">=0.3.34,<0.4.0"
typing-extensions = ">=4.13.2,<5.0.0"
graphviz = ">=0.20.3,<0.21.0"
python-dotenv = ">=1.1.0,<2.0.0"
requests = ">=2.31.0,<3.0.0"
fastapi = ">=0.110.0,<1.0.0"
uvicorn = ">=0.27.1,<1.0.0"
langchain = "^0.3.24"
langgraph-supervisor = "^0.0.20"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
